Various/MISP/requirements.txt
pymisp==2.5.10
PyYAML==6.0.2
requests==2.32.3
